refactor(Piece): 기물별 점수 반환 추상 메서드 생성
