+++ src/main/java/domain/piece/PieceScore.java
@@ -1,4 +1,4 @@
-package utils;
+package domain.piece;
 
 import java.util.Arrays;
